Move ScalarUdf to udf.rs
